pyproject.toml
[build-system]
requires = ["setuptools>=67", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "phishing-detection-suite"
version = "0.1.0"
description = "Benchmarking and SOC automation suite for phishing detection"
readme = "README.md"
authors = [{name = "OpenAI Codex"}]
license = {file = "LICENSE"}
requires-python = ">=3.10"
dependencies = [
    "beautifulsoup4>=4.12",
    "bs4>=0.0.1",
    "click>=8.1",
    "datasets>=2.16",
    "fastapi>=0.104",
    "httptools>=0.5",
    "httpx>=0.25",
    "jinja2>=3.1",
    "joblib>=1.3",
    "matplotlib>=3.8",
    "mlflow>=2.9",
    "numpy>=1.24",
    "pandas>=2.1",
    "python-dotenv>=1.0",
    "PyYAML>=6.0",
    "rich>=13.6",
    "scikit-learn>=1.3",
    "scipy>=1.11",
    "seaborn>=0.13",
    "tomli>=2.0",
    "toml>=0.10",
    "torch>=2.0",
    "transformers>=4.35",
    "tqdm>=4.66",
    "typer>=0.9",
    "uvicorn>=0.23",
    "xgboost>=1.7",
]

[project.optional-dependencies]
dev = [
    "black>=23.10",
    "mypy>=1.6",
    "pre-commit>=3.4",
    "pytest>=7.4",
    "pytest-asyncio>=0.21",
    "pytest-cov>=4.1",
    "ruff>=0.1",
    "types-PyYAML",
    "types-requests",
    "types-urllib3",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 88
target-version = ["py310"]

[tool.ruff]
line-length = 88
select = ["E", "F", "I", "B", "UP", "ANN", "ARG", "N", "C4", "T20", "PL", "RUF"]
ignore = ["ANN101", "ANN102", "ANN401", "D100", "D101", "D102", "D103", "D104", "D105"]
fix = false
src = ["src", "tests"]

[tool.ruff.isort]
known-first-party = ["common_utils", "task_a_benchmark", "task_b_soc"]

[tool.mypy]
python_version = "3.10"
strict = false
warn_unused_configs = true
warn_redundant_casts = true
warn_unused_ignores = true
show_error_codes = true
namespace_packages = true
mypy_path = ["src"]
plugins = []

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-m 'not slow'"

[tool.coverage.run]
branch = true
source = ["src"]
